replaced dummy events with fetched data
