+++ src/app/map/components/GoogleMaps/index.tsx
@@ -1,15 +1,15 @@
 "use client";
 
 import React, { useState, useEffect } from "react";
-import { useMantineColorScheme } from "@mantine/core";
+import { useMantineColorScheme, Image, Text } from "@mantine/core";
 import { mapTheme, loader } from "~/utils";
 import Calendar from "../Calendar";
 import EventCard from "../EventCard";
 import { useStyles } from "./styles";
 import { env } from "~/env.mjs";
 
-import { dummyData } from "../../event";
-import { Image, Text } from "@mantine/core";
+import { type ModifiedResult } from "~/app/api/events/all/route";
+import useFetchEvent from "~/hooks/useFetchEvent";
 import {
   GoogleMap,
   LoadScript,
@@ -38,8 +38,6 @@ const containerStyle = {
   width: "100vw",
   height: "100vh",
 };
-
-
 interface GoogleMapsProps {
   setMapLoaded: React.Dispatch<React.SetStateAction<boolean>>;
 }
@@ -48,10 +46,16 @@ export const GoogleMaps = ({ setMapLoaded }: GoogleMapsProps) => {
   const googleAPiKey = env.NEXT_PUBLIC_GOOGLE_MAPS_KEY;
   const { colorScheme } = useMantineColorScheme();
   const { classes } = useStyles();
-  const [infoWindowID, setInfoWindowID] = useState<number | null>(null);
+  const [date, setDate] = useState<Date | null>(new Date(Date.now()));
+  const { loading, error, result } = useFetchEvent<ModifiedResult[]>(
+    null,
+    date
+  );
+
+  const [infoWindowID, setInfoWindowID] = useState<string | null>(null);
   const [currentMarker, setCurrentMarker] = useState<JSX.Element | null>(null);
   const [markers, setMarkers] = useState<JSX.Element[] | null>(null); // Initialize as an empty array
-  const [date, setDate] = useState<Date | null>(new Date(Date.now()));
+
   const Loadings = (
     <div className={classes.gifWrapper}>
       <Image
@@ -71,49 +75,16 @@ export const GoogleMaps = ({ setMapLoaded }: GoogleMapsProps) => {
 
   useEffect(() => {
     const fetchMap = async () => {
+      setMapLoaded(true);
       await loader.load();
       navigator.geolocation.getCurrentPosition(
         (position) => {
           const { latitude, longitude } = position.coords;
-          const mapOptions = {
-            center: { lat: latitude, lng: longitude },
-            zoom: 10,
-            styles: colorScheme === "dark" ? mapTheme.dark : mapTheme.light,
-          };
           setCurrentMarker(
             <Marker position={{ lat: latitude, lng: longitude }} />
           );
 
-          if (dummyData !== null) {
-            const markerElements = dummyData.map((event, i) => {
-              const index = i + 1;
-              return (
-                event?.venue?.lat &&
-                event?.venue?.lng && (
-                  <Marker
-                    key={index}
-                    position={{
-                      lat: event.venue.lat,
-                      lng: event.venue.lng,
-                    }}
-                    onClick={() => {
-                      setInfoWindowID(index);
-                    }}
-                    icon={{ url: "/images/marker.svg" }}
-                  >
-                    {infoWindowID === index && (
-                      <InfoWindow>
-                        <EventCard event={event} />
-                      </InfoWindow>
-                    )}
-                  </Marker>
-                )
-              );
-            });
-            setMarkers(markerElements as JSX.Element[]);
-          }
-
-          setMapLoaded(true);
+          setMapLoaded(false);
         },
         (error) => {
           console.error("Error getting current location:", error);
@@ -121,8 +92,42 @@ export const GoogleMaps = ({ setMapLoaded }: GoogleMapsProps) => {
       );
     };
 
-    void fetchMap();
-  }, [colorScheme, setMapLoaded, infoWindowID]);
+    if (!loading && !error && result && result.length > 0) {
+      void fetchMap();
+    }
+  }, [colorScheme, setMapLoaded, loading, error, result]);
+
+  useEffect(() => {
+    if (!loading && !error && result && result.length > 0) {
+      const markerElements = result.map((event) => {
+        return (
+          event?.venue?.lat &&
+          event?.venue?.lng && (
+            <Marker
+              key={event.id}
+              position={{
+                lat: event.venue.lat,
+                lng: event.venue.lng,
+              }}
+              onClick={() => {
+                setInfoWindowID(event.id);
+              }}
+              icon={{ url: "/images/marker.svg" }}
+            >
+              {infoWindowID === event.id && (
+                <InfoWindow onCloseClick={() => setInfoWindowID(null)}>
+                  <EventCard event={event} />
+                </InfoWindow>
+              )}
+            </Marker>
+          )
+        );
+      });
+      setMarkers(markerElements as JSX.Element[]);
+    }
+  }, [loading, error, result, infoWindowID]);
+
+  console.log(infoWindowID);
 
   return (
     <div>
